% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstdev-package.r
\docType{package}
\name{Rstdev}
\alias{Rstdev}
\alias{Rstdev-package}
\title{Rstdev. calcualtes the standard deviation between two lists}
\description{
Rstdev. calcualtes the standard deviation between two lists
}

