% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstdev.R
\name{standard_deviation}
\alias{standard_deviation}
\title{Calculate standard deviation}
\usage{
standard_deviation(x)
}
\arguments{
\item{x}{list}
}
\value{
numeric
}
\description{
Calculates the standard deviation with a numeric vector or list of values
}
\examples{
standard_deviation(c(10,12,12,14))

}

