% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstdev.R
\name{standard_error}
\alias{standard_error}
\title{Calculate standard error}
\usage{
standard_error(x)
}
\arguments{
\item{x}{list}
}
\value{
numeric
}
\description{
Calculates the standard error with a numeric vector or list of values
}
\examples{
standard_error(c(10,12,12,14))

}

